// src/modules/vessel-trips/dto/export-vessel-trip.dto.ts
export class VesselTripTable {
  vessel: string;
  voyageId: string;
  imo: number;
  fromDate: string;
  toDate: string;
  distanceTraveled: number;
  hoursUnderway: string;
  mgo: number;
  lfo: number;
  hfo: number;
  vlsfo: number;
  lng: number;
  fuelCost: number;
  bunkerCost: number;
  freightCharges: number;
  originPort: string;
  destinationPort: string;
}